+++ s4e-web/src/app/errors/errors.model.ts
@@ -1,0 +1,7 @@
+export const HTTP_401_UNAUTHORIZED = 401;
+export const HTTP_403_FORBIDDEN = 403;
+export const HTTP_404_BAD_REQUEST = 400;
+export const HTTP_101_TIMEOUT = 101;
+export const HTTP_404_NOT_FOUND = 404;
+export const HTTP_500_INTERNAL_SERVER_ERROR = 500;
+export const HTTP_502_BAD_GATEWAY = 502;