+++ s4e-web/src/app/errors/errors.module.ts
@@ -1,0 +1,22 @@
+import {NgModule} from '@angular/core';
+import {CommonModule} from '@angular/common';
+import {ErrorsComponent} from './errors.component';
+import {RouterModule} from '@angular/router';
+
+@NgModule({
+  declarations: [ErrorsComponent],
+  imports: [
+    CommonModule,
+    RouterModule.forChild([
+      {
+        path: `errors/:errorCode`,
+        component: ErrorsComponent
+      }
+    ])
+  ],
+  exports: [
+    ErrorsComponent
+  ]
+})
+export class ErrorsModule {
+}