+++ s4e-web/src/app/utils/error-interceptor/error.interceptor.spec.ts
@@ -1,13 +1,45 @@
+import {RouterTestingModule} from '@angular/router/testing';
 import {ErrorInterceptor} from './error.interceptor';
+import {TestBed} from '@angular/core/testing';
+import {HttpClientTestingModule, HttpTestingController, TestRequest} from '@angular/common/http/testing';
+import {HTTP_INTERCEPTORS, HttpClient} from '@angular/common/http';
+import {Router} from '@angular/router';
+import {catchError} from 'rxjs/operators';
+import {of} from 'rxjs';
 
-describe('ErrorInterceptor', function () {
-  const interceptor = new ErrorInterceptor();
+describe('ErrorInterceptor', () => {
+  let errorInterceptor: ErrorInterceptor;
+  let httpController: HttpTestingController;
+  let http: HttpClient;
+  let router: Router;
 
-  it('should map to error.message', () => {
-    // :TODO add test
+  const url = `/api1/newTest`;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({
+      imports: [HttpClientTestingModule, RouterTestingModule],
+      providers: [
+        ErrorInterceptor,
+        {provide: HTTP_INTERCEPTORS, useClass: ErrorInterceptor, multi: true}
+      ]
+    });
+    router = TestBed.get(Router);
+    spyOn(router, 'navigate').and.stub();
+
+    errorInterceptor = TestBed.get(ErrorInterceptor);
+    httpController = TestBed.get(HttpTestingController);
+    http = TestBed.get(HttpClient);
   });
 
-  it('should map to statusText', () => {
-    // :TODO add test
+  it('Should handle client/server error', async () => {
+    const r = http.get(url).pipe(catchError(error => of(error))).toPromise();
+    const request: TestRequest = httpController.expectOne(url);
+    request.flush({}, {status: 500, statusText: 'Server Error'});
+
+    expect((await r).status).toBe(500);
+    expect((await r).message).toBe(`Http failure response for ${url}: 500 Server Error`);
+
+    httpController.verify();
+    expect(router.navigate).toHaveBeenCalled();
   });
 });