+++ s4e-web/src/app/views/map-view/state/view-configuration/view-configuration.service.ts
@@ -2,12 +2,12 @@ import {Injectable} from '@angular/core';
 import {HttpClient} from '@angular/common/http';
 import {ViewConfigurationStore} from './view-configuration.store';
 import {ViewConfiguration} from './view-configuration.model';
-import {catchError, finalize, map, shareReplay} from 'rxjs/operators';
+import {finalize, map, shareReplay} from 'rxjs/operators';
 import {S4eConfig} from '../../../../utils/initializer/config.service';
-import {Observable, throwError} from 'rxjs';
+import {Observable} from 'rxjs';
 import {IPageableResponse} from '../../../../state/pagable.model';
 import {ViewConfigurationQuery} from './view-configuration.query';
-import { catchErrorAndHandleStore } from 'src/app/common/store.util';
+import {catchErrorAndHandleStore} from '../../../../common/store.util';
 
 @Injectable({providedIn: 'root'})
 export class ViewConfigurationService {
@@ -21,6 +21,7 @@ export class ViewConfigurationService {
   add$(viewConfiguration: ViewConfiguration): Observable<boolean> {
     this.store.setError(null);
     this.store.setLoading(true);
+
     const r = this.http.post<ViewConfiguration>(`${this.config.apiPrefixV1}/saved-views`, viewConfiguration)
       .pipe(
         map(r => true),
@@ -36,11 +37,10 @@ export class ViewConfigurationService {
     this.store.setLoading(true);
     this.http.delete(`${this.config.apiPrefixV1}/saved-views/${uuid}`, {
       headers: {'Content-Type': 'application/json'}
-    })
-      .pipe(
-        catchErrorAndHandleStore(this.store),
-        finalize(() => this.store.setLoading(false))
-      )
+    }).pipe(
+      catchErrorAndHandleStore(this.store),
+      finalize(() => this.store.setLoading(false))
+    )
       .subscribe(() => this.store.remove(uuid));
   }
 
@@ -49,11 +49,9 @@ export class ViewConfigurationService {
     this.store.setLoading(true);
     this.http.get<IPageableResponse<ViewConfiguration>>(`${this.config.apiPrefixV1}/saved-views`, {
       headers: {'Content-Type': 'application/json'}
-    })
-      .pipe(
-        catchErrorAndHandleStore(this.store),
-        finalize(() => this.store.setLoading(false))
-      )
-      .subscribe(data => this.store.set(data.content));
+    }).pipe(
+      finalize(() => this.store.setLoading(false)),
+      catchErrorAndHandleStore(this.store),
+    ).subscribe(data => this.store.set(data.content));
   }
 }