+++ s4e-web/src/app/views/map-view/zk/configuration/state/configuration.service.ts
@@ -2,13 +2,12 @@ import {Injectable} from '@angular/core';
 import {HttpClient} from '@angular/common/http';
 import {ConfigurationStore} from './configuration.store';
 import {Configuration, ConfigurationState, ShareConfigurationRequest} from './configuration.model';
-import {finalize, map, shareReplay, catchError} from 'rxjs/operators';
+import {catchError, finalize, map, shareReplay} from 'rxjs/operators';
 import {ConfigurationQuery} from './configuration.query';
 import {S4eConfig} from '../../../../../utils/initializer/config.service';
 import {Dao} from '../../../../../common/dao.service';
-import {Observable, of} from 'rxjs';
 import {NotificationService} from 'notifications';
-import { catchErrorAndHandleStore } from 'src/app/common/store.util';
+import {Observable, of} from 'rxjs';
 
 @Injectable({providedIn: 'root'})
 export class ConfigurationService extends Dao<Configuration, ConfigurationState, ConfigurationStore> {
@@ -22,6 +21,7 @@ export class ConfigurationService extends Dao<Configuration, ConfigurationState,
 
   shareConfiguration(conf: ShareConfigurationRequest): Observable<boolean> {
     this.store.setLoading(true);
+
     const r = this.http.post<void>(`${this.config.apiPrefixV1}/share-link`, conf)
       .pipe(
         map(r => {